% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saipe_county}
\alias{saipe_county}
\title{County-Level Poverty Count and Rate Estimates}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 72259 rows and 21 columns.}
\usage{
saipe_county
}
\description{
Use \code{saipe::saipe_county}.
}
\keyword{datasets}
