% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saipe_state}
\alias{saipe_state}
\title{State-Level Poverty Count and Rate Estimates}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1122 rows and 20 columns.}
\usage{
saipe_state
}
\keyword{datasets}

