% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saipe_vars}
\alias{saipe_vars}
\title{SAIPE Variable Desciptions}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 18 rows and 5 columns.}
\usage{
saipe_vars
}
\keyword{datasets}

