% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saipe_us}
\alias{saipe_us}
\title{US Poverty Count and Rate Estimates}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 23 rows and 20 columns.}
\usage{
saipe_us
}
\description{
Use \code{saipe::saipe_us}.
}
\keyword{datasets}
